+++ backend/config/settings.py
@@ -1,8 +1,11 @@
-from pydantic import BaseSettings
+from pydantic_settings import BaseSettings
+from pydantic import ConfigDict
+import sys
 import os
 from dotenv import load_dotenv
 from pathlib import Path
 
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
 env_path = Path("/strapi/churchlink/.env")
 load_dotenv(dotenv_path=env_path)
 
@@ -14,9 +17,10 @@ class Settings(BaseSettings):
     STRAPI_API_TOKEN: str = os.getenv("API_TOKEN_SALT", "")
 
 
-    class Config:
-        env_file = str(env_path)
-        case_sensitive = True
+    model_config = ConfigDict(
+        env_file=str(env_path),
+        case_sensitive=True
+    )
 
 # Create settings instance
 settings = Settings()
