+++ backend/pyproject.toml
@@ -20,4 +20,6 @@ dependencies = [
     "faker>=37.1.0",
     "paypalrestsdk>=1.13.3",
     "pytest>=8.4.2",
+    "pydantic-settings>=2.11.0",
+    "windows-curses>=2.4.1",
 ]
