+++ backend/pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+testpaths = backend/tests
+python_files = *.py
+python_classes = Test*
+python_functions = test_*
+addopts = --rootdir=.. --import-mode=append
